Move Batch file staging implementation out to a separate assembly
