+++ src/Batch/Client/Batch.FileStaging/Azure.Batch.FileStaging.nuget.proj
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <ItemGroup>
+    <SdkNuGetPackage Include="Azure.Batch.FileStaging">
+      <PackageVersion>5.1.2</PackageVersion>
+      <Folder>$(MSBuildThisFileDirectory)</Folder>
+    </SdkNuGetPackage>
+  </ItemGroup>
+</Project>
