+++ src/Batch/Client/Batch.FileStaging/ErrorMessages.resx
