+++ src/Batch/Client/Batch.FileStaging/SequentialFileStaging.cs
