+++ src/Batch/Client/Batch.FileStaging/SequentialFileStagingArtifact.cs
