+++ src/Batch/Client/Batch.FileStaging/StagingStorageAccount.cs
