+++ src/Batch/Client/Src/Batch.csproj
@@ -17,19 +17,6 @@
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Microsoft.CSharp" />
-    <Reference Include="Microsoft.Data.Edm">
-      <HintPath>$(LibraryNugetPackageFolder)\Microsoft.Data.Edm.5.6.4\lib\net40\Microsoft.Data.Edm.dll</HintPath>
-    </Reference>
-    <Reference Include="Microsoft.Data.OData">
-      <HintPath>$(LibraryNugetPackageFolder)\Microsoft.Data.OData.5.6.4\lib\net40\Microsoft.Data.OData.dll</HintPath>
-    </Reference>
-    <Reference Include="Microsoft.Data.Services.Client">
-      <HintPath>$(LibraryNugetPackageFolder)\Microsoft.Data.Services.Client.5.6.4\lib\net40\Microsoft.Data.Services.Client.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
-    <Reference Include="Microsoft.Azure.KeyVault.Core">
-      <HintPath>$(LibraryNugetPackageFolder)\Microsoft.Azure.KeyVault.Core.1.0.0\lib\portable-net45+wp8+wpa81+win\Microsoft.Azure.KeyVault.Core.dll</HintPath>
-    </Reference>
     <Reference Include="Microsoft.Rest.ClientRuntime">
       <HintPath>$(LibraryNugetPackageFolder)\Microsoft.Rest.ClientRuntime.2.3.1\lib\net45\Microsoft.Rest.ClientRuntime.dll</HintPath>
       <Private>True</Private>
@@ -37,9 +24,6 @@
     <Reference Include="Microsoft.Rest.ClientRuntime.Azure">
       <HintPath>$(LibraryNugetPackageFolder)\Microsoft.Rest.ClientRuntime.Azure.3.3.1\lib\net45\Microsoft.Rest.ClientRuntime.Azure.dll</HintPath>
       <Private>True</Private>
-    </Reference>
-    <Reference Include="Microsoft.WindowsAzure.Storage">
-      <HintPath>$(LibraryNugetPackageFolder)\WindowsAzure.Storage.6.2.0\lib\net40\Microsoft.WindowsAzure.Storage.dll</HintPath>
     </Reference>
     <Reference Include="Newtonsoft.Json">
       <HintPath>$(LibraryNugetPackageFolder)\Newtonsoft.Json.6.0.8\lib\net45\Newtonsoft.Json.dll</HintPath>
@@ -52,10 +36,6 @@
     <Reference Include="System.Net.Http" />
     <Reference Include="System.Net.Http.WebRequest" />
     <Reference Include="System.Runtime.Serialization" />
-    <Reference Include="System.Spatial">
-      <HintPath>$(LibraryNugetPackageFolder)\System.Spatial.5.6.4\lib\net40\System.Spatial.dll</HintPath>
-      <Private>True</Private>
-    </Reference>
     <Reference Include="System.Web" />
     <Reference Include="System.Xml.Linq" />
     <Reference Include="System.Data" />
@@ -71,6 +51,7 @@
     <Compile Include="Common\Enumerations\OnAllTasksComplete.cs" />
     <Compile Include="Common\Enumerations\OSType.cs" />
     <Compile Include="Common\Enumerations\AddTaskStatus.cs" />
+    <Compile Include="FileStaging\FileStagingUtils.cs" />
     <Compile Include="ImageReference.cs" />
     <Compile Include="IPropertyMetadata.cs" />
     <Compile Include="IReadOnly.cs" />
@@ -113,11 +94,6 @@
     <Compile Include="CloudJob.cs" />
     <Compile Include="Common\BatchErrorCodeStrings.cs" />
     <Compile Include="DetailLevel.cs" />
-    <Compile Include="FileStaging\ErrorMessages.Designer.cs">
-      <DependentUpon>ErrorMessages.resx</DependentUpon>
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-    </Compile>
     <Compile Include="IInheritedBehaviors.cs" />
     <Compile Include="IPagedEnumerable.cs" />
     <Compile Include="IPagedEnumerator.cs" />
@@ -240,19 +216,8 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FileStaging\StagingStorageAccount.cs" />
     <Compile Include="FileStaging\Constants.cs" />
-    <Compile Include="FileStaging\SequentialFileStaging.cs" />
-    <Compile Include="FileStaging\FileStagingManager.cs" />
-    <Compile Include="FileStaging\FileStagingLinkedSources.cs" />
-    <Compile Include="FileStaging\IFileStagingManager.cs" />
     <Compile Include="FileStaging\IFileStagingProvider.cs" />
-    <Compile Include="FileStaging\SequentialFileStagingArtifact.cs" />
-    <EmbeddedResource Include="FileStaging\ErrorMessages.resx">
-      <SubType>Designer</SubType>
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>ErrorMessages.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
