+++ src/Batch/Client/Tests/ObjectModel/Azure.Batch.Unit.Tests/StaticAnalysisUnitTests.cs
@@ -122,7 +122,6 @@
                      "ArgumentOutOfRangeException",
                      "InvalidOperationException",
                      "ArgumentException",
-                     "FileNotFoundException",
                      "ParallelOperationsException",
                  };
